Set image.tag in the lacework-agent-5.4.2 charts (#55)
The tarball is the source of truth and it had an empty
value for '.image.tag'. Running this command resulted
in a diff compared with the source files checked into git:

    tar -xf lacework-agent-5.4.2.tgz

To fix the tarball, I ran this command from a clean checkout:

    tar -cf lacework-agent-5.4.2.tgz lacework-agent

I then re-ran the 'helm repo index' step as well. I have verified
that unpacking the 5.4.2 tarball no longer results in unstaged
changes.