Lacework Agent: Support pre-release version of GKE
GKE uses pre-release version to identify Kubernetes version. Unfortunately,
this does not work with the compatability version that we added as part of
OpenShift support, so our Helm Charts refuse to install the agent now that we
specify kubeVersion. See helm/helm#3810 for more discussion on the topic.

By adding `-0` to the end of kubeVersion, we tell Helm that pre-release
Kubernetes versions are okay.