Agent Charts: Enable custom keys at top-level
Give customers the ability to mix-in their own custom values files that
may contain keys unknown to the Lacework Agent charts. The tradeoff is
that our charts will continue to install even if the customer provides a
typo'd value name, however that should be rare for values specified at
the top-level.